# server/docs/swagger.yaml
definitions:
...
  request.AdminCreateProject:
    properties:
      description:
        type: string
      name:
        type: string
      owner_id:
        type: integer
    required:
    - description
    - name
    - owner_id
    type: object
  request.AdminRegister:
    properties:
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - name
    - password
    type: object
  request.AdminUpdatePassword:
    properties:
      password:
        minLength: 8
        type: string
      user_id:
        type: integer
    required:
    - password
    - user_id
    type: object
  request.ManualMessage:
    properties:
      content:
        type: string
      link:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    required:
    - content
    - user_ids
    type: object
  request.MessageRead:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  request.MessageShareLink:
    properties:
      link:
        type: string
      to_user_id:
        type: integer
    required:
    - link
    - to_user_id
    type: object
...
  response.CommonResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.CommonResponseWithoutData:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/create_project:
    post:
      consumes:
      - application/json
      description: admin create_project
      parameters:
      - description: AdminCreateProject
        in: body
        name: AdminCreateProject
        required: true
        schema:
          $ref: '#/definitions/request.AdminCreateProject'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: admin create_project
      tags:
      - Admin
  /admin/register:
    post:
      consumes:
      - application/json
      description: admin register
      parameters:
      - description: AdminRegister
        in: body
        name: AdminRegister
        required: true
        schema:
          $ref: '#/definitions/request.AdminRegister'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: admin register
      tags:
      - Admin
...
  /admin/update_password:
    post:
      consumes:
      - application/json
      description: admin update_password
      parameters:
      - description: AdminUpdatePassword
        in: body
        name: AdminUpdatePassword
        required: true
        schema:
          $ref: '#/definitions/request.AdminUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: admin update_password
      tags:
      - Admin
  /message/manual:
    post:
      consumes:
      - application/json
      description: message manual
      parameters:
      - description: MessageRead
        in: body
        name: MessageRead
        required: true
        schema:
          $ref: '#/definitions/request.ManualMessage'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: message manual
      tags:
      - Message
  /message/read:
    post:
      consumes:
      - application/json
      description: message read
      parameters:
      - description: MessageRead
        in: body
        name: MessageRead
        required: true
        schema:
          $ref: '#/definitions/request.MessageRead'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: message read
      tags:
      - Message
  /message/receive_list:
    get:
      consumes:
      - '*/*'
      description: message receive_list
      parameters:
      - default: 1
        in: query
        name: page
        type: integer
      - default: 10
        in: query
        name: page_size
        type: integer
      - in: query
        name: read
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: message receive_list
      tags:
      - Message
  /message/send_list:
    get:
      consumes:
      - '*/*'
      description: message send_list
      parameters:
      - default: 1
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 10
        in: query
        minimum: 5
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: message send_list
      tags:
      - Message
  /message/share_link:
    post:
      consumes:
      - application/json
      description: message share link
      parameters:
      - description: MessageShareLink
        in: body
        name: MessageShareLink
        required: true
        schema:
          $ref: '#/definitions/request.MessageShareLink'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: message share link
      tags:
      - Message
...
swagger: "2.0"